Fix substancial performance bottleneck fon ParagraphBuilderFactory
