+++ package/cpp/api/JsiSkParagraphBuilder.h
@@ -113,7 +113,7 @@ public:
                                  sk_sp<SkFontMgr> fontManager)
       : JsiSkHostObject(std::move(context)) {
     _fontCollection = sk_make_sp<para::FontCollection>();
-    _fontCollection->setDefaultFontManager(getContext()->createFontMgr());
+    _fontCollection->setDefaultFontManager(JsiSkFontMgrFactory::getFontMgr(context));
     if (fontManager != nullptr) {
       _fontCollection->setAssetFontManager(fontManager);
     }
@@ -125,35 +125,4 @@ private:
   std::unique_ptr<para::ParagraphBuilder> _builder;
   sk_sp<para::FontCollection> _fontCollection;
 };
-
-/**
- Implementation of the ParagraphBuilderFactory for making ParagraphBuilder JSI
- object
- */
-class JsiSkParagraphBuilderFactory : public JsiSkHostObject {
-public:
-  JSI_HOST_FUNCTION(Make) {
-    // Get paragraph style from params
-    auto paragraphStyle =
-        count > 0 ? JsiSkParagraphStyle::fromValue(runtime, arguments[0])
-                  : para::ParagraphStyle();
-
-    // get font manager
-    auto fontMgr =
-        count > 1 ? JsiSkTypefaceFontProvider::fromValue(runtime, arguments[1])
-                  : nullptr;
-
-    // Create the paragraph builder
-    return jsi::Object::createFromHostObject(
-        runtime, std::make_shared<JsiSkParagraphBuilder>(
-                     getContext(), paragraphStyle, fontMgr));
-  }
-
-  JSI_EXPORT_FUNCTIONS(JSI_EXPORT_FUNC(JsiSkParagraphBuilderFactory, Make))
-
-  explicit JsiSkParagraphBuilderFactory(
-      std::shared_ptr<RNSkPlatformContext> context)
-      : JsiSkHostObject(std::move(context)) {}
-};
-
 } // namespace RNSkia
